// src/utils/ads/medialpha/utils/types.ts
export type MediaalphaInlineAdType ={
    _6_1_svg_image_url: string;
    ad_id: number;
    buyer: string;
    click_url: string;
    delivery: string;
    description: string;
    display_url: string;
    half_medium_image_url: string;
    half_small_image_url: string;
    half_svg_image_url: string;
    headline: string;
    iframe_support: number;
    impr_html: string;
    large_image_url: string;
    medium_image_url: string;
    no_iframe: number;
    small_image_url: string;
    svg_image_url: string;
  }


  